perf: ExecuteInterfaceTrait is added to implement the interface execute method
